ISSUE #1738 - Move model icon to form header
