+++ frontend/routes/viewerGui/components/issues/components/issueDetails/issueDetails.styles.ts
@@ -17,7 +17,7 @@
 
 import styled from 'styled-components';
 import { LogList as LogListBase } from '../../../../../components/logList/logList.component';
-import { PreviewDetails as PreviewDetailsBase } from '../../../previewDetails/previewDetails.component';
+import PreviewDetailsBase from '../../../previewDetails/previewDetails.container';
 
 export const Container = styled.div``;
